Can now change input base

+++ Float64.h
@@ -24,8 +24,8 @@ class f64 : public Printable
   // member variable (the number)
   float64_t        num_;
   static char      str_[];
-  static int8_t   aft_point;
-  static int8_t   obase;
+  static int   aft_point;
+  static int   obase;
 
   void init(void);
 
@@ -60,8 +60,8 @@ public:
   //operator double () const;
   //operator float32_t () const;
   operator float64_t () const;
-  void setDecs(int8_t n); /* set no of digits after the decimal point */
-  void setBase(int8_t n); /* arbitrary base selection */
+  void setDecs(int n); /* set no of digits after the decimal point */
+  void setBase(int n); /* arbitrary base selection */
   virtual size_t printTo(Print& p) const; // for Serial.print()
 
   // operators ... assignment
